Support manylinux_2_24

+++ tests/test_manylinux_utils_docker.sh
@@ -3,8 +3,9 @@
 source manylinux_utils.sh
 source tests/utils.sh
 
-suppress yum_install rsync && suppress yum erase -y rsync \
-    || ingest "yum_install valid"
-suppress bash -c '! yum_install nonexistent' || ingest "yum_install nonexistent"
-
-barf
+if [[ $MB_ML_VER != "_2_24" ]]; then
+    suppress yum_install rsync && suppress yum erase -y rsync \
+        || ingest "yum_install valid"
+    suppress bash -c '! yum_install nonexistent' || ingest "yum_install nonexistent"
+    barf
+fi
